Move editor to component + v-model
