+++ examples/vue/src/ui/blockSideMenuFactory.ts
@@ -1,5 +1,5 @@
-import type { BlockSideMenuFactory } from "@blocknote/core";
-import { createApp } from 'vue';
+import type { BlockSideMenuFactory } from "@blocknote/core"
+import { createApp } from 'vue'
 
 import BlockSideMenu from '@/components/BlockSideMenu.vue'
 
@@ -8,25 +8,25 @@ import BlockSideMenu from '@/components/BlockSideMenu.vue'
  * It renders a drag handle and + button to create a new block
  */
 export const blockSideMenuFactory: BlockSideMenuFactory = (staticParams) => {
-  
-    const container = document.createElement("div");
-    const instance = createApp(BlockSideMenu, {
-      staticParams
-    }).mount(container);
-    document.body.appendChild(instance.$el);
 
-    return {
-      element: instance.$el,
-      render: (params, isHidden) => {
-        if (isHidden) {
-          instance.$el.style.display = "block";
-        }
+  const container = document.createElement("div")
+  const instance = createApp(BlockSideMenu, {
+    staticParams
+  }).mount(container)
+  document.body.appendChild(instance.$el)
 
-        instance.$el.style.top = params.referenceRect.y + "px";
-        instance.$el.style.left = params.referenceRect.x - instance.$el.offsetWidth + "px";
-      },
-      hide: () => {
-        instance.$el.style.display = "none";
-      },
-    };
+  return {
+    element: instance.$el,
+    render: (params, isHidden) => {
+      if (isHidden) {
+        instance.$el.style.display = "block"
+      }
+
+      instance.$el.style.top = params.referenceRect.y + "px"
+      instance.$el.style.left = params.referenceRect.x - instance.$el.offsetWidth + "px"
+    },
+    hide: () => {
+      instance.$el.style.display = "none"
+    },
   }
+}
